requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
yookassa==3.0.0
openai==1.12.0
requests==2.31.0
pydantic==2.5.3
fastapi==0.108.0
uvicorn==0.25.0
pandas==2.1.4
schedule==1.2.0